Separate global and local domain config
